Feitas alterações de impressão, excluidos construtores que não eram utilizados.

+++ src/entities/Average.java
@@ -1,7 +1,4 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+
 package entities;
 
 import interfaces.CalculateAverage;
@@ -10,7 +7,7 @@ import interfaces.CalculateAverage;
 public class Average implements CalculateAverage{
 
     @Override
-    public double calculateAverage(double oldKm, double currentKm, double liters) {
+    public float calculateAverage(float oldKm, float currentKm, float liters) {
        return (currentKm - oldKm) / liters; 
     }
