+++ src/interfaces/CalculateAverage.java
@@ -1,15 +1,8 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
+
 package interfaces;
 
-/**
- *
- * @author vinicius-dev
- */
 public interface CalculateAverage {
     
-    public double calculateAverage(double km, double currentKm, double liters);
+    public float calculateAverage(float km, float currentKm, float liters);
     
 }
